reworked breakpoints
git-svn-id: https://kdlibcpp.svn.codeplex.com/svn@89541 9b283d60-5439-405e-af05-b73fd8c4d996

+++ kdlib/include/kdlib/dbgtypedef.h
@@ -10,7 +10,7 @@ typedef unsigned long           PROCESS_ID;
 typedef unsigned long           PROCESS_DEBUG_ID;
 typedef unsigned long           THREAD_ID;
 typedef unsigned long           THREAD_DEBUG_ID;
-typedef unsigned long long      BREAKPOINT_ID;
+typedef unsigned long           BREAKPOINT_ID;
 typedef unsigned long long      MEMOFFSET_64;
 typedef unsigned long           MEMOFFSET_32;
 typedef long                    MEMOFFSET_REL;
